*VG 041310: cif-file/src/ParentChild.C Purpose: Fixing compilation error on GCC 3.2 Changes: - Added static casting of iterator distance in order to fix compilation error.

+++ src/ParentChild.C
@@ -402,7 +402,8 @@ void ParentChild::ISTableFindPairs(
             for (multimap<string, string, StringLess>::iterator it =
               range.first; it != range.second; ++it)
             {
-                if (distance(range.first, it) == orderNum)
+                if (static_cast<unsigned int>(distance(range.first, it)) ==
+                  orderNum)
                 {
                     childKeys.push_back((*it).second);
                 }
